src/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
prometheus-client==0.21.0
python-dotenv==1.0.1
pytest==8.3.3
sqlite3  # Built-in with Python